chore: Fix circular dependency and missing return errors in the web-component-analyzer package
