+++ packages/web-component-analyzer/src/analyze/flavors/js-doc/discover-global-features.ts
@@ -13,25 +13,30 @@ export const discoverGlobalFeatures: AnalyzerFlavor["discoverGlobalFeatures"] = 
 		if (context.ts.isInterfaceDeclaration(node) && node.name.text === "HTMLElement") {
 			return discoverFeatures.csspart?.(node, context);
 		}
+		return;
 	},
 	cssproperty: (node: Node, context: AnalyzerVisitContext): ComponentCssProperty[] | undefined => {
 		if (context.ts.isInterfaceDeclaration(node) && node.name.text === "HTMLElement") {
 			return discoverFeatures.cssproperty?.(node, context);
 		}
+		return;
 	},
 	event: (node: Node, context: AnalyzerVisitContext): ComponentEvent[] | undefined => {
 		if (context.ts.isInterfaceDeclaration(node) && node.name.text === "HTMLElement") {
 			return discoverFeatures.event?.(node, context);
 		}
+		return;
 	},
 	slot: (node: Node, context: AnalyzerVisitContext): ComponentSlot[] | undefined => {
 		if (context.ts.isInterfaceDeclaration(node) && node.name.text === "HTMLElement") {
 			return discoverFeatures.slot?.(node, context);
 		}
+		return;
 	},
 	member: (node: Node, context: AnalyzerVisitContext): ComponentMember[] | undefined => {
 		if (context.ts.isInterfaceDeclaration(node) && node.name.text === "HTMLElement") {
 			return discoverFeatures?.member?.(node, context);
 		}
+		return;
 	}
 };
