+++ packages/web-component-analyzer/src/analyze/flavors/jsx/discover-global-features.ts
@@ -37,5 +37,7 @@ export const discoverGlobalFeatures: AnalyzerFlavor["discoverGlobalFeatures"] = 
 
 			return members;
 		}
+
+		return;
 	}
 };
