+++ packages/web-component-analyzer/src/analyze/flavors/lwc/discover-members.ts
@@ -25,6 +25,8 @@ export function discoverMembers(node: Node, context: AnalyzerDeclarationVisitCon
 	if (ts.isSetAccessor(node) || ts.isGetAccessor(node) || ts.isPropertyDeclaration(node) || ts.isPropertySignature(node)) {
 		return parsePropertyDecorator(node, context);
 	}
+
+	return;
 }
 
 /**
