+++ packages/web-component-analyzer/src/analyze/stages/analyze-declaration.ts
@@ -2,8 +2,8 @@ import { Node } from "typescript";
 import { AnalyzerVisitContext } from "../analyzer-visit-context";
 import { AnalyzerDeclarationVisitContext, ComponentFeatureCollection } from "../flavors/analyzer-flavor";
 import { ComponentDeclaration } from "../types/component-declaration";
-import { getNodeName, getSymbol, resolveDeclarations } from "../util/ast-util";
 import { getJsDoc } from "../util/js-doc-util";
+import { getSymbol, resolveDeclarations, getNodeName } from "../util/resolve-declarations";
 import { discoverFeatures } from "./discover-features";
 import { discoverInheritance } from "./discover-inheritance";
 import { excludeNode } from "./flavor/exclude-node";
