+++ packages/web-component-analyzer/src/analyze/stages/discover-definitions.ts
@@ -2,7 +2,7 @@ import { Node, SourceFile } from "typescript";
 import { AnalyzerVisitContext } from "../analyzer-visit-context";
 import { ComponentDeclaration } from "../types/component-declaration";
 import { ComponentDefinition } from "../types/component-definition";
-import { getSymbol, resolveSymbolDeclarations } from "../util/ast-util";
+import { getSymbol, resolveSymbolDeclarations } from "../util/resolve-declarations";
 import { visitDefinitions } from "./flavor/visit-definitions";
 
 /**
