+++ packages/web-component-analyzer/src/transformers/json2/json2-transformer.ts
@@ -6,9 +6,10 @@ import { AnalyzerResult } from "../../analyze/types/analyzer-result";
 import { ComponentDeclaration, ComponentHeritageClause } from "../../analyze/types/component-declaration";
 import { ComponentFeatureBase } from "../../analyze/types/features/component-feature";
 import { JsDoc } from "../../analyze/types/js-doc";
-import { findParent, getNodeName, resolveDeclarations } from "../../analyze/util/ast-util";
+import { findParent } from "../../analyze/util/ast-util";
 import { getMixinHeritageClauses, getSuperclassHeritageClause, visitAllHeritageClauses } from "../../analyze/util/component-declaration-util";
 import { getJsDoc } from "../../analyze/util/js-doc-util";
+import { resolveDeclarations, getNodeName } from "../../analyze/util/resolve-declarations";
 import { arrayDefined } from "../../util/array-util";
 import { getTypeHintFromMethod } from "../../util/get-type-hint-from-method";
 import { getTypeHintFromType } from "../../util/get-type-hint-from-type";
